fix(e2e): keep the Vue hyperlink test hermetic and move the Vue source link to shared
Enrolling the Vue interaction spec brought the suite's only real external
navigation with it. The link's target is now stubbed on the browser context, so
the popup resolves without egress and a runner with no internet reports a
hyperlink failure only when hyperlinks are broken.

`SourceLink.vue` joins its React twin and the other cross-example parts in
`examples/shared`, where the widened lint and format globs already reach it.
