+++ examples/vue/src/EditorChrome.vue
@@ -114,7 +114,7 @@ import {
 import { blankDocumentBytes } from '@docx-editor.dev/core/editor';
 import BrandLogo from '../../shared/BrandLogo.vue';
 import AdapterSwitcher from './AdapterSwitcher.vue';
-import SourceLink from './SourceLink.vue';
+import SourceLink from '../../shared/SourceLink.vue';
 import ThemeToggle from './ThemeToggle.vue';
 import FontPreviewItems from './FontPreviewItems.vue';
 import { DEMO_PRIMARY_BUTTON, DEMO_SECONDARY_BUTTON, keepCaret } from './demoButtons';
